add retry package, fix fieldlabel class, add button to go to satis homepage and date of last build, split code between server/client/lib
