+++ client/satisEasy.css
@@ -1,0 +1,7 @@
+.container {
+    margin-top: 70px;
+}
+
+.nav-wrapper .left li {
+    margin-left: 5px;
+}
