--- satisEasy.css
@@ -1,3 +1,0 @@
-.container {
-    margin-top: 70px;
-}
